COMMERCE-1453 Refactor CPDataSource -> CPHttpDataSource and remove suppression
