+++ commerce-product-content-web/src/main/java/com/liferay/commerce/product/content/web/internal/display/context/CPPublisherConfigurationDisplayContext.java
@@ -25,8 +25,8 @@ import com.liferay.commerce.product.catalog.CPSku;
 import com.liferay.commerce.product.content.render.list.CPContentListRendererRegistry;
 import com.liferay.commerce.product.content.render.list.entry.CPContentListEntryRendererRegistry;
 import com.liferay.commerce.product.content.web.internal.util.CPPublisherWebHelper;
-import com.liferay.commerce.product.data.source.CPDataSource;
 import com.liferay.commerce.product.data.source.CPDataSourceRegistry;
+import com.liferay.commerce.product.data.source.CPHttpDataSource;
 import com.liferay.commerce.product.item.selector.criterion.CPDefinitionItemSelectorCriterion;
 import com.liferay.commerce.product.type.CPTypeServicesTracker;
 import com.liferay.commerce.product.util.CPInstanceHelper;
@@ -241,8 +241,8 @@ public class CPPublisherConfigurationDisplayContext
 		return null;
 	}
 
-	public List<CPDataSource> getCPDataSources() {
-		return _cpDataSourceRegistry.getCPDataSources();
+	public List<CPHttpDataSource> getCPHttpDataSources() {
+		return _cpDataSourceRegistry.getCPHttpDataSources();
 	}
 
 	public CPSku getDefaultCPSku(CPCatalogEntry cpCatalogEntry)
