+++ commerce-product-content-web/src/main/resources/META-INF/resources/init.jsp
@@ -55,8 +55,8 @@ page import="com.liferay.commerce.product.content.web.internal.portlet.CPCompare
 page import="com.liferay.commerce.product.content.web.internal.portlet.CPCompareContentPortlet" %><%@
 page import="com.liferay.commerce.product.content.web.internal.portlet.CPContentPortlet" %><%@
 page import="com.liferay.commerce.product.content.web.internal.portlet.CPPublisherPortlet" %><%@
-page import="com.liferay.commerce.product.data.source.CPDataSource" %><%@
 page import="com.liferay.commerce.product.data.source.CPDataSourceResult" %><%@
+page import="com.liferay.commerce.product.data.source.CPHttpDataSource" %><%@
 page import="com.liferay.commerce.product.model.CPInstance" %><%@
 page import="com.liferay.commerce.product.model.CPOptionCategory" %><%@
 page import="com.liferay.commerce.product.model.CPSpecificationOption" %><%@
