+++ commerce-product-service/src/main/java/com/liferay/commerce/product/internal/data/source/CPHttpDataSourceAssetCategoriesImpl.java
@@ -15,17 +15,15 @@
 package com.liferay.commerce.product.internal.data.source;
 
 import com.liferay.asset.kernel.model.AssetEntry;
-import com.liferay.asset.kernel.model.AssetTag;
 import com.liferay.asset.kernel.service.AssetEntryLocalService;
 import com.liferay.commerce.product.catalog.CPQuery;
-import com.liferay.commerce.product.data.source.CPDataSource;
+import com.liferay.commerce.product.data.source.CPHttpDataSource;
 import com.liferay.commerce.product.model.CPDefinition;
 import com.liferay.commerce.product.util.CPDefinitionHelper;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.Portal;
 
-import java.util.List;
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
@@ -36,17 +34,18 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true,
-	property = "commerce.product.data.source.name=" + CPDataSourceAssetTagsImpl.NAME,
-	service = CPDataSource.class
+	property = "commerce.product.data.source.name=" + CPHttpDataSourceAssetCategoriesImpl.NAME,
+	service = CPHttpDataSource.class
 )
-public class CPDataSourceAssetTagsImpl extends BaseCPDataSourceAssetEntryImpl {
+public class CPHttpDataSourceAssetCategoriesImpl
+	extends BaseCPHttpDataSourceAssetEntryImpl {
 
-	public static final String NAME = "assetTagsDataSource";
+	public static final String NAME = "assetCategoriesDataSource";
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			getResourceBundle(locale), "products-of-the-same-tags");
+			getResourceBundle(locale), "products-of-the-same-categories");
 	}
 
 	@Override
@@ -58,26 +57,12 @@ public class CPDataSourceAssetTagsImpl extends BaseCPDataSourceAssetEntryImpl {
 	protected CPQuery getCPQuery(long cpDefinitionId) throws PortalException {
 		CPQuery cpQuery = new CPQuery();
 
-		cpQuery.setAnyTagIds(_getTagIds(cpDefinitionId));
-
-		return cpQuery;
-	}
-
-	private long[] _getTagIds(long cpDefinitionId) throws PortalException {
 		AssetEntry assetEntry = _assetEntryLocalService.getEntry(
 			CPDefinition.class.getName(), cpDefinitionId);
 
-		List<AssetTag> assetTags = assetEntry.getTags();
+		cpQuery.setAnyCategoryIds(assetEntry.getCategoryIds());
 
-		long[] tagIds = new long[assetTags.size()];
-
-		for (int i = 0; i < assetTags.size(); i++) {
-			AssetTag assetTag = assetTags.get(i);
-
-			tagIds[i] = assetTag.getTagId();
-		}
-
-		return tagIds;
+		return cpQuery;
 	}
 
 	@Reference(unbind = "-")
